requirements.txt
langchain
langchain-community
langgraph
fastapi 
uvicorn
pydantic
pymongo[srv]
python-dotenv
langchain-google-genai